fix: add custom hook for sparkle app url
